-修正RederField的FieldType=Double时出现的客户端JS错误（零度九分.拍豆-8936）。 	+修正DataKeyNames值是小数（比如9.80），则页面回发时会出现F_STATE验证出错的问题（Fire-8910）。 		-JavaScript会自动忽略小数的最后零（比如2.0,9.80），导致服务器生成的数据和回发的数据不一致。 	-修正后台更新复选框列表后无法获取选中项的问题（究级闪电EX-8926）。
git-svn-id: https://fineui.svn.codeplex.com/svn@98150 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ FineUI_v6/FineUI/Business/Enums/Grid/FieldType.cs
@@ -65,8 +65,10 @@ namespace FineUI
                 case FieldType.Boolean:
                     result = "boolean";
                     break;
+                    // ExtJS v6.0不支持客户端指定 double 字符串，需要改为 float
+                    // 客户端处理一样，这个设置只影响服务器端的解析
                 case FieldType.Double:
-                    result = "double";
+                    result = "float";
                     break;
                 case FieldType.Date:
                     result = "date";
